Updates for merge with latest main, ensuring related tests complete ok
