+++ tests/operators/test_scoreswrappers.py
@@ -78,7 +78,7 @@ def test_scores_rmse_nonzero(cube: iris.cube.Cube):
 
 def test_scores_rmse_no_time_coord(cube):
     """RMSE of cubes with no time coordinate."""
-    c1 = cube.extract(iris.Constraint(time=datetime.datetime(2022, 9, 21, 3, 30)))
+    c1 = cube.extract(iris.Constraint(time=datetime.datetime(2022, 9, 21, 3, 00)))
     c1.remove_coord("time")
     c2 = c1.copy()
     del c2.attributes["cset_comparison_base"]
